Move QuadItems to its own file
